\name{OMEXDIAparms}
\docType{data}
\alias{OMEXDIAparms}
\title{OMEXDIA diagenetic model parameters}
\description{a vector with parameters of the OMEXDIA diagenetic model}
\usage{OMEXDIAparms}
\format{a vector with the names and values of the OMEXDIA diagenetic model parameters.
See Soetaert et al., 1996 for further details}
\author{Karline Soetaert <k.soetaert@nioo.knaw.nl>}
\examples{OMEXDIAparms
OMEXDIAparms["rFast"]
as.list(OMEXDIAparms)$rFast}
\references{Soetaert K, PMJ Herman and JJ Middelburg, 1996a. 
A model of early diagenetic processes from the shelf to abyssal depths. 
Geochimica Cosmochimica Acta, 60(6):1019-1040.}
\seealso{\code{\link{OMEXDIAmodel}}}
\note{To be used as input to \code{\link{OMEXDIAmodel}}}
\keyword{datasets}

